consolidate validating into flux aggregator
